Hide private methods, use absolute imports, modularize tests.
