+++ xarray_jax/tests/xarray_jax_test.py
@@ -306,269 +306,6 @@ class XarrayJaxTest(absltest.TestCase):
     }
     chex.assert_trees_all_equal(outputs_dict, inputs_and_inputs_coords_dict)
 
-  def test_flatten_unflatten_variable(self):
-    variable = xarray.Variable(
-        ('lat', 'lon'), jnp.ones((3, 4), dtype=np.float32))
-    children, aux = xarray_jax._flatten_variable(variable)
-    # Check auxiliary info is hashable/comparable (important for jax.jit):
-    hash(aux)
-    self.assertEqual(aux, aux)
-    roundtrip = xarray_jax._unflatten_variable(aux, children)
-    self.assertTrue(variable.equals(roundtrip))
-
-  def test_flatten_unflatten_data_array(self):
-    data_array = xarray_jax.DataArray(
-        data=jnp.ones((3, 4), dtype=np.float32),
-        dims=('lat', 'lon'),
-        coords={'lat': np.arange(3)},
-        jax_coords={'lon': np.arange(4) * 10},
-    )
-    children, aux = xarray_jax._flatten_data_array(data_array)
-    # Check auxiliary info is hashable/comparable (important for jax.jit):
-    hash(aux)
-    self.assertEqual(aux, aux)
-    roundtrip = xarray_jax._unflatten_data_array(aux, children)
-    self.assertTrue(data_array.equals(roundtrip))
-
-  def test_flatten_unflatten_dataset(self):
-    foo = jnp.ones((3, 4), dtype=np.float32)
-    bar = jnp.ones((2, 3, 4), dtype=np.float32)
-    dataset = xarray_jax.Dataset(
-        data_vars={'foo': (('lat', 'lon'), foo),
-                   'bar': (('time', 'lat', 'lon'), bar)},
-        coords={
-            'time': np.arange(2),
-            'lat': np.arange(3) * 10},
-        jax_coords={
-            'lon': np.arange(4) * 10})
-    children, aux = xarray_jax._flatten_dataset(dataset)
-    # Check auxiliary info is hashable/comparable (important for jax.jit):
-    hash(aux)
-    self.assertEqual(aux, aux)
-    roundtrip = xarray_jax._unflatten_dataset(aux, children)
-    self.assertTrue(dataset.equals(roundtrip))
-
-  def test_flatten_unflatten_datatree(self):
-    # Coords to be inherited from the parent dataset, we include one jax
-    # coord and one not to check both code paths
-    parent_dataset = xarray_jax.Dataset(
-        jax_coords={'time': xarray.Variable(('time',), np.arange(2))},
-        coords={'lon': xarray.Variable(('lon',), np.arange(4) * 10)})
-
-    bar = jnp.ones((2, 3, 4), dtype=np.float32)
-    child_dataset = xarray_jax.Dataset(
-        {'bar': (('time', 'lat', 'lon'), bar)},
-        coords={'lat': np.arange(3)})
-
-    datatree = xarray.DataTree(
-        dataset=parent_dataset,
-        children={'child': xarray.DataTree(dataset=child_dataset)})
-
-    children, aux = xarray_jax._flatten_datatree(datatree)
-    # Check auxiliary info is hashable/comparable (important for jax.jit):
-    hash(aux)
-    self.assertEqual(aux, aux)
-    roundtrip = xarray_jax._unflatten_datatree(aux, children)
-    self.assertTrue(datatree.equals(roundtrip))
-
-  def test_flatten_unflatten_added_dim(self):
-    data_array = xarray_jax.DataArray(
-        data=jnp.ones((3, 4), dtype=np.float32),
-        dims=('lat', 'lon'),
-        coords={'lat': np.arange(3),
-                'lon': np.arange(4) * 10})
-    leaves, treedef = jax.tree_util.tree_flatten(data_array)
-    leaves = [jnp.expand_dims(x, 0) for x in leaves]
-    with xarray_jax.dims_change_on_unflatten(lambda dims: ('new',) + dims):
-      with_new_dim = jax.tree_util.tree_unflatten(treedef, leaves)
-    self.assertEqual(('new', 'lat', 'lon'), with_new_dim.dims)
-    xarray.testing.assert_identical(
-        jax.device_get(data_array),
-        jax.device_get(with_new_dim.isel(new=0)))
-
-  def test_map_added_dim(self):
-    data_array = xarray_jax.DataArray(
-        data=jnp.ones((3, 4), dtype=np.float32),
-        dims=('lat', 'lon'),
-        coords={'lat': np.arange(3),
-                'lon': np.arange(4) * 10})
-    with xarray_jax.dims_change_on_unflatten(lambda dims: ('new',) + dims):
-      with_new_dim = jax.tree_util.tree_map(lambda x: jnp.expand_dims(x, 0),
-                                            data_array)
-    self.assertEqual(('new', 'lat', 'lon'), with_new_dim.dims)
-    xarray.testing.assert_identical(
-        jax.device_get(data_array),
-        jax.device_get(with_new_dim.isel(new=0)))
-
-  def test_map_remove_dim(self):
-    foo = jnp.ones((1, 3, 4), dtype=np.float32)
-    bar = jnp.ones((1, 2, 3, 4), dtype=np.float32)
-    dataset = xarray_jax.Dataset(
-        data_vars={'foo': (('batch', 'lat', 'lon'), foo),
-                   'bar': (('batch', 'time', 'lat', 'lon'), bar)},
-        coords={
-            'batch': np.array([123]),
-            'time': np.arange(2),
-            'lat': np.arange(3) * 10,
-            'lon': np.arange(4) * 10})
-    with xarray_jax.dims_change_on_unflatten(lambda dims: dims[1:]):
-      with_removed_dim = jax.tree_util.tree_map(lambda x: jnp.squeeze(x, 0),
-                                                dataset)
-    self.assertEqual(('lat', 'lon'), with_removed_dim['foo'].dims)
-    self.assertEqual(('time', 'lat', 'lon'), with_removed_dim['bar'].dims)
-    self.assertNotIn('batch', with_removed_dim.dims)
-    self.assertNotIn('batch', with_removed_dim.coords)
-    xarray.testing.assert_identical(
-        jax.device_get(dataset.isel(batch=0, drop=True)),
-        jax.device_get(with_removed_dim))
-
-  def test_pmap(self):
-    devices = jax.local_device_count()
-    foo = jnp.zeros((devices, 3, 4), dtype=np.float32)
-    bar = jnp.zeros((devices, 2, 3, 4), dtype=np.float32)
-    dataset = xarray_jax.Dataset({
-        'foo': (('device', 'lat', 'lon'), foo),
-        'bar': (('device', 'time', 'lat', 'lon'), bar)})
-
-    def func(d):
-      self.assertNotIn('device', d.dims)
-      return d + 1
-    func = xarray_jax.pmap(func, dim='device')
-
-    result = func(dataset)
-    xarray.testing.assert_identical(
-        jax.device_get(dataset + 1),
-        jax.device_get(result))
-
-    # Can call it again with a different argument structure (it will recompile
-    # under the hood but should work):
-    dataset = dataset.drop_vars('foo')
-    result = func(dataset)
-    xarray.testing.assert_identical(
-        jax.device_get(dataset + 1),
-        jax.device_get(result))
-
-  def test_pmap_with_jax_coords(self):
-    devices = jax.local_device_count()
-    foo = jnp.zeros((devices, 3, 4), dtype=np.float32)
-    bar = jnp.zeros((devices, 2, 3, 4), dtype=np.float32)
-    time = jnp.zeros((devices, 2), dtype=np.float32)
-    dataset = xarray_jax.Dataset(
-        {'foo': (('device', 'lat', 'lon'), foo),
-         'bar': (('device', 'time', 'lat', 'lon'), bar)},
-        coords={
-            'lat': np.arange(3),
-            'lon': np.arange(4),
-        },
-        jax_coords={
-            # Currently any jax_coords need a leading device dimension to use
-            # with pmap, same as for data_vars.
-            # TODO(matthjw): have pmap automatically broadcast to all devices
-            # where the device dimension not present.
-            'time': xarray.Variable(('device', 'time'), time),
-        }
-    )
-
-    def func(d):
-      self.assertNotIn('device', d.dims)
-      self.assertNotIn('device', d.coords['time'].dims)
-
-      # The jax_coord 'time' should be passed in backed by a JAX array, but
-      # not as an index coordinate.
-      self.assertIsInstance(d.coords['time'].data, jax.Array)
-      self.assertNotIn('time', d.indexes)
-
-      return d + 1
-    func = xarray_jax.pmap(func, dim='device')
-
-    result = func(dataset)
-    xarray.testing.assert_identical(
-        jax.device_get(dataset + 1),
-        jax.device_get(result))
-
-    # Can call it again with a different argument structure (it will recompile
-    # under the hood but should work):
-    dataset = dataset.drop_vars('foo')
-    result = func(dataset)
-    xarray.testing.assert_identical(
-        jax.device_get(dataset + 1),
-        jax.device_get(result))
-
-  def test_pmap_with_tree_mix_of_xarray_and_jax_array(self):
-    devices = jax.local_device_count()
-    data_array = xarray_jax.DataArray(
-        data=jnp.ones((devices, 3, 4), dtype=np.float32),
-        dims=('device', 'lat', 'lon'))
-    plain_array = jnp.ones((devices, 2), dtype=np.float32)
-    inputs = {'foo': data_array,
-              'bar': plain_array}
-
-    def func(x):
-      return x['foo'] + 1, x['bar'] + 1
-
-    func = xarray_jax.pmap(func, dim='device')
-    result_foo, result_bar = func(inputs)
-    xarray.testing.assert_identical(
-        jax.device_get(inputs['foo'] + 1),
-        jax.device_get(result_foo))
-    np.testing.assert_array_equal(
-        jax.device_get(inputs['bar'] + 1),
-        jax.device_get(result_bar))
-
-  def test_pmap_complains_when_dim_not_first(self):
-    devices = jax.local_device_count()
-    data_array = xarray_jax.DataArray(
-        data=jnp.ones((3, devices, 4), dtype=np.float32),
-        dims=('lat', 'device', 'lon'))
-
-    func = xarray_jax.pmap(lambda x: x+1, dim='device')
-    with self.assertRaisesRegex(
-        ValueError, 'Expected dim device at index 0, found at 1'):
-      func(data_array)
-
-  def test_apply_ufunc(self):
-    inputs = xarray_jax.DataArray(
-        data=jnp.asarray([[1, 2], [3, 4]]),
-        dims=('x', 'y'),
-        coords={'x': [0, 1],
-                'y': [2, 3]})
-    result = xarray_jax.apply_ufunc(
-        lambda x: jnp.sum(x, axis=-1),
-        inputs,
-        input_core_dims=[['x']])
-    expected_result = xarray_jax.DataArray(
-        data=[4, 6],
-        dims=('y',),
-        coords={'y': [2, 3]})
-    xarray.testing.assert_identical(expected_result, jax.device_get(result))
-
-  def test_apply_ufunc_multiple_return_values(self):
-    def ufunc(array):
-      return jnp.min(array, axis=-1), jnp.max(array, axis=-1)
-    inputs = xarray_jax.DataArray(
-        data=jnp.asarray([[1, 4], [3, 2]]),
-        dims=('x', 'y'),
-        coords={'x': [0, 1],
-                'y': [2, 3]})
-    result = xarray_jax.apply_ufunc(
-        ufunc, inputs, input_core_dims=[['x']], output_core_dims=[[], []])
-    expected = (
-        # Mins:
-        xarray_jax.DataArray(
-            data=[1, 2],
-            dims=('y',),
-            coords={'y': [2, 3]}
-        ),
-        # Maxes:
-        xarray_jax.DataArray(
-            data=[3, 4],
-            dims=('y',),
-            coords={'y': [2, 3]}
-        )
-    )
-    xarray.testing.assert_identical(expected[0], jax.device_get(result[0]))
-    xarray.testing.assert_identical(expected[1], jax.device_get(result[1]))
-
   def test_eval_shape_with_xarray(self):
     # This needs jax.ShapeDtypeStruct to be wrappable inside xarray types.
     function = jax.jit(lambda v: v + 1)
@@ -578,63 +315,6 @@ class XarrayJaxTest(absltest.TestCase):
     self.assertIsInstance(output_shapes, xarray.Variable)
     self.assertEqual(output_shapes.shape, (3, 4))
     self.assertEqual(output_shapes.dtype, np.float32)
-
-  def test_scan(self):
-    def f(carry, x):
-      dataset_input, plain_jax_input = x
-      self.assertEqual(plain_jax_input.shape, ())
-      self.assertEqual(dataset_input.sizes, {'x': 2})  # No 'time' dimension.
-      carry = carry + 1
-      y = dataset_input + carry
-      return carry, y
-
-    dataset_inputs = xarray_jax.Dataset(
-        data_vars={
-            # Put the scan dimension (time) second to make sure it transposes
-            # the array appropriately for JAX which requires scan dim first.
-            'foo': (('x', 'time'), jnp.zeros((2, 5))),
-        },
-        coords={
-            'x': np.arange(2),
-            'time': np.arange(5) * 10,
-        },
-        jax_coords={
-            'time_extra': (('time',), np.arange(5)*2),
-        })
-    # These must have scan dimension first:
-    plain_jax_inputs = jnp.zeros((5,))
-
-    carry, result = xarray_jax.scan(f,
-                                    init=0,
-                                    xs=(dataset_inputs, plain_jax_inputs),
-                                    dim='time')
-    self.assertEqual(jax.device_get(carry), 5)
-    self.assertEqual(result.foo.sizes, {'x': 2, 'time': 5})
-    self.assertIn('x', result.coords)
-    # Unfortunately static coordinates along the dimension we scan over will
-    # not be preserved on the result:
-    self.assertNotIn('time', result.coords)
-    # The jax_coord along time will still be present though:
-    np.testing.assert_array_equal(
-        jax.device_get(result.foo.coords['time_extra'].data),
-        jax.device_get(dataset_inputs.coords['time_extra'].data))
-
-  def test_scan_no_inputs(self):
-    def f(carry, x):
-      assert x is None
-      carry = carry + 1
-      y = carry * 10
-      # Set a jax_coord on the output to check it maps through to the result
-      # correctly:
-      y = xarray_jax.assign_jax_coords(y, extra_coord=carry[0])
-      return carry, y
-
-    init = xarray_jax.DataArray(data=np.zeros(2), dims=('x',))
-    carry, result = xarray_jax.scan(f, init=init, length=5, dim='time')
-    np.testing.assert_array_equal(jax.device_get(carry), [5, 5])
-    self.assertEqual(result.sizes, {'x': 2, 'time': 5})
-    np.testing.assert_array_equal(jax.device_get(result.extra_coord.data),
-                                  [1, 2, 3, 4, 5])
 
   def test_assign_coords_arg_types(self):
     # Check we can assign coords with a variety of shorthands mirroring those
